requirements.txt
Flask
Flask-SQLAlchemy
python-dotenv
google-genai